Move Containerd component into containerd subpackage
The subpackage already contains the configurer. Place the component next
to it. It's now accessible as containerd.NewComponent(...) instead of
worker.NewContainerd(...).
