+++ pkg/component/worker/containerd/utils_windows.go
@@ -13,7 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package worker
+
+package containerd
 
 import (
 	"os"
